moved class ScenarioExecutorImpl and interface ScenarioExecutor to package maintenance
