+++ executor-service/src/main/java/executor/service/ExecutionService.java
@@ -1,5 +1,6 @@
 package executor.service;
 
+import executor.service.maintenance.plugin.proxy.ScenarioExecutor;
 import org.openqa.selenium.WebDriver;
 
 public interface ExecutionService {
