+++ executor-service/src/main/java/executor/service/ParallelFlowExecutorImpl.java
@@ -5,6 +5,7 @@ import executor.service.factory.webdriverinitializer.WebDriverProvider;
 
 import executor.service.maintenance.plugin.proxy.ProxySourcesClient;
 
+import executor.service.maintenance.plugin.proxy.ScenarioExecutor;
 import executor.service.model.ProxyConfigHolderDto;
 import executor.service.model.ThreadPoolConfigDto;
