+++ executor-service/src/main/java/executor/service/maintenance/plugin/proxy/ScenarioExecutor.java
@@ -1,4 +1,4 @@
-package executor.service;
+package executor.service.maintenance.plugin.proxy;
 
 import executor.service.model.ScenarioDto;
 import org.openqa.selenium.WebDriver;
