+++ executor-service/src/test/java/executor/service/ExecutionServiceImplTest.java
@@ -1,5 +1,7 @@
 package executor.service;
 
+import executor.service.maintenance.plugin.proxy.ScenarioExecutor;
+import executor.service.maintenance.plugin.proxy.ScenarioExecutorImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
