+++ executor-service/src/test/java/executor/service/ScenarioExecutorImplTest.java
@@ -1,5 +1,6 @@
 package executor.service;
 
+import executor.service.maintenance.plugin.proxy.ScenarioExecutorImpl;
 import executor.service.model.ScenarioDto;
 import executor.service.stepexecution.ClickCss;
 import executor.service.stepexecution.ClickXpath;
